Proto: Split data formats from DataSource.proto into new DataFormat.proto
