+++ common-test/src/main/java/feast/common/it/DataGenerator.java
@@ -20,16 +20,16 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.Duration;
 import com.google.protobuf.Timestamp;
+import feast.proto.core.DataFormatProto.FileFormat;
+import feast.proto.core.DataFormatProto.FileFormat.ParquetFormat;
+import feast.proto.core.DataFormatProto.StreamFormat;
+import feast.proto.core.DataFormatProto.StreamFormat.AvroFormat;
+import feast.proto.core.DataFormatProto.StreamFormat.ProtoFormat;
 import feast.proto.core.DataSourceProto.DataSource;
 import feast.proto.core.DataSourceProto.DataSource.BigQueryOptions;
 import feast.proto.core.DataSourceProto.DataSource.FileOptions;
 import feast.proto.core.DataSourceProto.DataSource.KafkaOptions;
 import feast.proto.core.DataSourceProto.DataSource.KinesisOptions;
-import feast.proto.core.DataSourceProto.FileFormat;
-import feast.proto.core.DataSourceProto.FileFormat.ParquetFormat;
-import feast.proto.core.DataSourceProto.StreamFormat;
-import feast.proto.core.DataSourceProto.StreamFormat.AvroFormat;
-import feast.proto.core.DataSourceProto.StreamFormat.ProtoFormat;
 import feast.proto.core.EntityProto;
 import feast.proto.core.FeatureProto;
 import feast.proto.core.FeatureProto.FeatureSpecV2;
