+++ core/src/main/java/feast/core/model/DataSource.java
@@ -21,14 +21,14 @@ import com.google.protobuf.Message;
 import com.google.protobuf.MessageOrBuilder;
 import com.google.protobuf.util.JsonFormat;
 import feast.core.util.TypeConversion;
+import feast.proto.core.DataFormatProto.FileFormat;
+import feast.proto.core.DataFormatProto.StreamFormat;
 import feast.proto.core.DataSourceProto;
 import feast.proto.core.DataSourceProto.DataSource.BigQueryOptions;
 import feast.proto.core.DataSourceProto.DataSource.FileOptions;
 import feast.proto.core.DataSourceProto.DataSource.KafkaOptions;
 import feast.proto.core.DataSourceProto.DataSource.KinesisOptions;
 import feast.proto.core.DataSourceProto.DataSource.SourceType;
-import feast.proto.core.DataSourceProto.FileFormat;
-import feast.proto.core.DataSourceProto.StreamFormat;
 import java.util.HashMap;
 import java.util.Map;
 import javax.persistence.Column;
